# :zap: faster parallel searchsorted