This PR aims at optimizing the runtime of parallel downsampling with x, resolves #14 

To do so, the following optimizations are applied;
- :lock_with_ink_pen: no pre-allocation of output array (with mutex) -> (0.02s => 0.008s)
- :thread: `get_equidistant_bin_idx_iterator_parallel` now creates a nested iterator -> (0.008 -> 0.005s)
  - :exploding_head: the outer iterator yield `# threads` iterators
    => this allows to re-use the left index to perform a more efficient search sorted procedure (as is done in the sequential Iterator - see description in #14)